feat：Remove `EssentialFeed` module import from `FeedViewController` file since it does not depend on any `EssentialFeed` component.
